Add more tests

+++ src/main/kotlin/de/roamingthings/workbench/sealed/Animal.kt
@@ -16,5 +16,5 @@ data class Cat(
 
 data class Fish(
     val finCount: Int,
-    val size: Float
+    val size: Double
 ) : Animal()
